requirements.txt
django>=2.0
markdown
Pygments
mysqlclient
python-dateutil
django-tz-detect
django-ckeditor